2003-06-26  Al Riddoch  <alriddoch@zepler.org>
	* app/BladeMap.cpp, app/EntityLayer.cpp, app/HeightManager.cpp,
	  app/Holo.cpp, app/IsoMap.cpp, app/Server.cpp,
	  app/ServerEntities.cpp, equator_plugin.cpp:
	  Clean up some compiler warnings.

	* app/IsoMap.cpp, app/Texture.cpp, app/Texture.h:
	  Migrate tile code into separate file from texture so that
	  the tile and texture code is unified with apogee.

+++ app/Holo.cpp
@@ -16,7 +16,7 @@ void Holo::draw(GlView & view)
     float winsize = std::max(view.get_width(), view.get_height());
 
     float winscale = view.getScale();
-    int numlines = winsize / (40.0f * winscale) + 1;
+    int numlines = (int)(winsize / (40.0f * winscale) + 1);
     int xc = -(int)view.getXoff();
     int yc = -(int)view.getYoff();
